Bringing linux dev up to date with the latest mainline.

Conflicts:
	src/cinder/Url.cpp
	src/cinder/Utilities.cpp

+++ src/cinder/app/App.cpp
@@ -439,7 +439,7 @@ fs::path App::getOpenFilePath( const fs::path &initialPath, vector<string> exten
 	if( ! initialPath.empty() )
 		[cinderOpen setDirectoryURL:[NSURL fileURLWithPath:[[NSString stringWithUTF8String:initialPath.c_str()] stringByExpandingTildeInPath]]];
 
-	int resultCode = [cinderOpen runModal];
+	NSInteger resultCode = [cinderOpen runModal];
 
 	setFullScreen( wasFullScreen );
 	restoreWindowContext();
@@ -480,7 +480,7 @@ fs::path App::getFolderPath( const fs::path &initialPath )
 	if( ! initialPath.empty() )
 		[cinderOpen setDirectoryURL:[NSURL fileURLWithPath:[[NSString stringWithUTF8String:initialPath.c_str()] stringByExpandingTildeInPath]]];
 	
-	int resultCode = [cinderOpen runModal];	
+	NSInteger resultCode = [cinderOpen runModal];
 	
 	setFullScreen(wasFullScreen);
 	restoreWindowContext();
@@ -542,7 +542,7 @@ fs::path App::getSaveFilePath( const fs::path &initialPath, vector<string> exten
 			[cinderSave setNameFieldStringValue:file];
 	}
 	
-	int resultCode = [cinderSave runModal];
+	NSInteger resultCode = [cinderSave runModal];
 
 	setFullScreen( wasFullScreen );
 	restoreWindowContext();
